requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.14.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
comtypes==1.4.14
contourpy==1.3.2
cycler==0.12.1
DateTime==5.4
decorator==5.2.1
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
docopt==0.6.2
filelock==3.19.1
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
g4f==0.6.2.9
geocoder==1.38.1
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
Js2Py==0.74
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
multidict==6.6.4
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.2.7
nest-asyncio==1.6.0
numpu==0.1.4
numpy==2.2.6
opencv-python==4.13.0.92
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pipwin==0.5.2
propcache==0.3.2
protobuf==3.20.3
PyAudio==0.2.14
pycryptodome==3.23.0
pyjsparser==2.7.1
pyparsing==3.2.5
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyttsx3==2.99
pytz==2023.3.post1
pywin32==311
PyYAML==6.0.2
ratelim==0.1.6
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
soupsieve==2.8.1
SpeechRecognition==3.14.4
sqlparse==0.5.5
tabulate==0.9.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.6
wikipedia==1.4.0
yarl==1.20.1
zope.interface==6.1
